backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
celery>=5.3.0,<6.0.0
redis>=4.5.0,<5.0.0
psycopg2-binary>=2.9.0,<3.0.0
requests>=2.31.0,<3.0.0
PyYAML>=6.0,<7.0
